+++ app/src/main/java/dnd/jon/spellbook/CharacterSpellEntry.java
@@ -11,13 +11,13 @@ import androidx.room.Index;
         )
 class CharacterSpellEntry {
 
-    @ColumnInfo(name = "character_id") final int characterID;
-    @ColumnInfo(name = "spell_id") final int spellID;
+    @ColumnInfo(name = "character_id") final long characterID;
+    @ColumnInfo(name = "spell_id") final long spellID;
     @ColumnInfo(name = "favorite", defaultValue = "0") final boolean favorite;
     @ColumnInfo(name = "known", defaultValue = "0") final boolean known;
     @ColumnInfo(name = "prepared", defaultValue = "0") final boolean prepared;
 
-    CharacterSpellEntry(int characterID, int spellID, boolean favorite, boolean known, boolean prepared) {
+    CharacterSpellEntry(long characterID, long spellID, boolean favorite, boolean known, boolean prepared) {
         this.characterID = characterID;
         this.spellID = spellID;
         this.favorite = favorite;
